change score from int to float

+++ Assets/Code/Application/Modules/Game/Commands/CalculateScore/CalculateScoreCommand.cs
@@ -3,7 +3,7 @@ using Code.Application.Commons.Interfaces.Mediator;
 
 namespace Code.Application.Modules.Game.Commands.CalculateScore
 {
-    public class CalculateScoreCommand : ICommand<int>
+    public class CalculateScoreCommand : ICommand<float>
     {
         public float StartTime { get; set; }
         public float EndTime { get; set; }
